modules/hybrid_search/requirements.txt
opensearch-py>=2.6.0
qdrant-client>=1.9.0
sentence-transformers>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
